refactor: moving the devtestlabs client
