+++ azurerm/internal/provider/services.go
@@ -18,6 +18,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/datafactory"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/datalake"
 	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/devspace"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/internal/services/devtestlabs"
 )
 
 func SupportedServices() []common.ServiceRegistration {
@@ -38,5 +39,6 @@ func SupportedServices() []common.ServiceRegistration {
 		datafactory.Registration{},
 		datalake.Registration{},
 		devspace.Registration{},
+		devtestlabs.Registration{},
 	}
 }
